app/requirements.txt
flask
mysql-connector==2.2.9